terraform_data sensitive and write-only tools
Add a new object to the terraform_data resource to handle sensitive and
write-only attributes.

The new `store` object has input and output values, which work much like
the top-level input and output. The difference is that the `store.input`
is truly "write only" and can also accept ephemeral values.

Additional arguments for the `store` object are:
- `version`, allows the user to determine exactly when the input
  value will be replaced in `store.output`.
- `sensitive`, conditionally directs the stored value between `output`
  or `sensitive_output`.
- `replace` signals that any change to either output should trigger
  replacement of the entire resource.